CTS-52 fix: move frame function to windowing.py
