+++ caits/windowing.py
@@ -126,6 +126,36 @@ def windowing_df(
         return "Invalid mode. Use 'dict' or 'df' as mode."
 
 
+def frame_signal(
+        array: np.ndarray,
+        frame_length: int,
+        hop_length: int
+) -> np.ndarray:
+    """Distinguishes a signal into overlapping frames.
+
+    Args:
+        array: The input signal as a numpy.ndarray.
+        frame_length: The length of the frame in samples.
+        hop_length: The number of samples to advance between frames (overlap).
+
+    Returns:
+        numpy.ndarray: The framed signal in the form of a 2D numpy.ndarray
+            (frame_length x num_frames).
+    """
+    # Number of frames
+    num_frames = 1 + int(np.floor((len(array) - frame_length) / hop_length))
+    # Row indices
+    rows = np.arange(frame_length)[:, None]
+    # Column indices
+    cols = np.arange(num_frames) * hop_length
+    # Index matrix for each frame
+    indices = rows + cols
+    # Frame the signal according to calculated indices
+    frames = array[indices]
+    return frames
+
+
+
 def create_chunks(
         array: np.ndarray,
         chunk_length: int
